move QA components to new QA folder, update routes accordingly
