+++ client/src/components/QA/components/QAnswer.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 class QAnswer extends React.Component {
   constructor(props) {
