+++ client/src/components/QA/components/QAsk.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 class QAsk extends React.Component {
   constructor(props) {
